requirements.txt
streamlit==1.11.1
sumatra-client
sklearn
matplotlib